♻️ refactor(routing): swap intro and editor page entrypoints
- move the editor implementation out of the root app component into a dedicated editor page
- repurpose the root app component as the intro landing screen
- update route mapping so / renders intro and /editor renders the editor page
- replace the pages barrel export from intro to editor and remove the legacy intro page file
- bump @jbpark/ui-kit to ^3.3.0 and refresh the lockfile accordingly
