+++ package.json
@@ -83,7 +83,7 @@
     "@dnd-kit/modifiers": "^9.0.0",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@jbpark/ui-kit": "^3.2.1",
+    "@jbpark/ui-kit": "^3.3.0",
     "@jbpark/use-hooks": "^2.7.0",
     "@tailwindcss/vite": "^4.1.13",
     "@tiptap/core": "^3.22.5",
