+++ src/main.tsx
@@ -3,14 +3,14 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import App from './App';
-import { Intro, Playground, Preview } from './pages';
+import { Editor, Playground, Preview } from './pages';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Intro />} />
-        <Route path="/editor" element={<App />} />
+        <Route path="/" element={<App />} />
+        <Route path="/editor" element={<Editor />} />
         <Route path="/preview" element={<Preview />} />
         <Route path="/playground" element={<Playground />} />
       </Routes>
