+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-export { default as Intro } from './intro';
+export { default as Editor } from './editor';
 export { default as Playground } from './playground';
 export { default as Preview } from './preview';
